Adding controller and service
